+++ app/services/subscription/decode_jwt_token.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+class Subscription::DecodeJwtToken
+  def initialize(token)
+    @token = token
+  end
+
+  def call
+    JWT.decode(
+      @token,
+      ENV['JWT_SECRET_KEY'],
+      true,
+      { algorithm: 'HS256' }
+    ).first.symbolize_keys
+  end
+end