+++ spec/services/imports/create_spec.rb
@@ -17,8 +17,8 @@ RSpec.describe Imports::Create do
                            content_type: 'application/json')
       end
 
-      it 'calls the GoogleMaps::SemanticHistoryParser' do
-        expect(GoogleMaps::SemanticHistoryParser).to \
+      it 'calls the GoogleMaps::SemanticHistoryImporter' do
+        expect(GoogleMaps::SemanticHistoryImporter).to \
           receive(:new).with(import, user.id).and_return(double(call: true))
         service.call
       end
@@ -31,15 +31,15 @@ RSpec.describe Imports::Create do
     context 'when source is google_phone_takeout' do
       let(:import) { create(:import, source: 'google_phone_takeout') }
 
-      it 'calls the GoogleMaps::PhoneTakeoutParser' do
-        expect(GoogleMaps::PhoneTakeoutParser).to \
+      it 'calls the GoogleMaps::PhoneTakeoutImporter' do
+        expect(GoogleMaps::PhoneTakeoutImporter).to \
           receive(:new).with(import, user.id).and_return(double(call: true))
         service.call
       end
     end
 
     context 'when source is owntracks' do
-      let(:import) { create(:import, source: 'owntracks') }
+      let(:import) { create(:import, source: 'owntracks', name: '2024-03.rec') }
       let(:file_path) { Rails.root.join('spec/fixtures/files/owntracks/2024-03.rec') }
       let(:file) { Rack::Test::UploadedFile.new(file_path, 'application/octet-stream') }
 
@@ -54,12 +54,6 @@ RSpec.describe Imports::Create do
       end
 
       context 'when import is successful' do
-        it 'creates a finished notification' do
-          service.call
-
-          expect(user.notifications.last.kind).to eq('info')
-        end
-
         it 'schedules stats creating' do
           Sidekiq::Testing.inline! do
             expect { service.call }.to \
@@ -79,10 +73,38 @@ RSpec.describe Imports::Create do
           allow(OwnTracks::Importer).to receive(:new).with(import, user.id).and_raise(StandardError)
         end
 
-        it 'creates a failed notification' do
-          service.call
+        context 'when self-hosted' do
+          before do
+            allow(DawarichSettings).to receive(:self_hosted?).and_return(true)
+          end
 
-          expect(user.notifications.last.kind).to eq('error')
+          after do
+            allow(DawarichSettings).to receive(:self_hosted?).and_call_original
+          end
+
+          it 'creates a failed notification' do
+            service.call
+
+            expect(user.notifications.last.content).to \
+              include('Import "2024-03.rec" failed: StandardError, stacktrace: ')
+          end
+        end
+
+        context 'when not self-hosted' do
+          before do
+            allow(DawarichSettings).to receive(:self_hosted?).and_return(false)
+          end
+
+          after do
+            allow(DawarichSettings).to receive(:self_hosted?).and_call_original
+          end
+
+          it 'does not create a failed notification' do
+            service.call
+
+            expect(user.notifications.last.content).to \
+              include('Import "2024-03.rec" failed, please contact us at hi@dawarich.com')
+          end
         end
       end
     end
@@ -107,8 +129,8 @@ RSpec.describe Imports::Create do
     context 'when source is geojson' do
       let(:import) { create(:import, source: 'geojson') }
 
-      it 'calls the Geojson::ImportParser' do
-        expect(Geojson::ImportParser).to \
+      it 'calls the Geojson::Importer' do
+        expect(Geojson::Importer).to \
           receive(:new).with(import, user.id).and_return(double(call: true))
         service.call
       end
@@ -117,8 +139,8 @@ RSpec.describe Imports::Create do
     context 'when source is immich_api' do
       let(:import) { create(:import, source: 'immich_api') }
 
-      it 'calls the Photos::ImportParser' do
-        expect(Photos::ImportParser).to \
+      it 'calls the Photos::Importer' do
+        expect(Photos::Importer).to \
           receive(:new).with(import, user.id).and_return(double(call: true))
         service.call
       end
@@ -127,8 +149,8 @@ RSpec.describe Imports::Create do
     context 'when source is photoprism_api' do
       let(:import) { create(:import, source: 'photoprism_api') }
 
-      it 'calls the Photos::ImportParser' do
-        expect(Photos::ImportParser).to \
+      it 'calls the Photos::Importer' do
+        expect(Photos::Importer).to \
           receive(:new).with(import, user.id).and_return(double(call: true))
         service.call
       end