+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+# 0.25.6 - 2025-04-23
+
+## Changed
+
+- Import edit page now allows to edit import name.
+- Importing data now does not create a notification for the user.
+- Updating stats now does not create a notification for the user.
+
+## Fixed
+
+- Fixed a bug where an import was failing due to partial file download. #1069 #1073 #1024 #1051
+
 # 0.25.5 - 2025-04-18
 
 This release introduces a new way to send transactional emails using SMTP. Example may include password reset, email confirmation, etc.